% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdviz-package.R
\docType{package}
\name{hdviz-package}
\alias{hdviz}
\alias{hdviz-package}
\title{hdviz: Homodatum Viz}
\description{
Vector definitions for abstract data types
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/datasketch/hdviz}
  \item Report bugs at \url{https://github.com/datasketch/hdviz/issues}
}

}
\author{
\strong{Maintainer}: Juan Pablo Marin Diaz \email{jpmarindiaz@gmail.com}

}
\keyword{internal}
