% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdviz.R
\name{hdviz}
\alias{hdviz}
\title{Create a hdviz object}
\usage{
hdviz(
  viz,
  name = NULL,
  description = NULL,
  slug = NULL,
  meta = NULL,
  width = NULL,
  height = NULL,
  formats = NULL,
  ...
)
}
\arguments{
\item{name}{a custom name can be added}

\item{description}{a custom description can be added}

\item{slug}{a custom slug can be added. If not, hdtable will try creating one.}

\item{meta}{Custom Metadata can be added}

\item{x}{A data frame}

\item{hdtable_type}{The type of hdtable to create}
}
\value{
A hdtable object
}
\description{
Create a hdtable object from a data frame. The main value of a hdtable is its metadata. When creating it, hdtable will add to the data frame the following information:
\itemize{
\item viz: visualization object (ggplot or htmlwidget)
\item name: Name for the hdviz, setted on \emph{name} argument
\item description: Description for the hdviz, setted on \emph{description} argument
\item slug: a custom slug can be added
}
}
\examples{
hdtable(mtcars, hdtable_type = "Num", name = "MTCars")

}
