% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdviz.R
\name{is_hdviz}
\alias{is_hdviz}
\title{is_hdviz}
\usage{
is_hdviz(x)
}
\arguments{
\item{x}{object to be coerced or tested}
}
\value{
returns TRUE or FALSE depending on whether its argument is of type hdviz or not.
}
\description{
test for objects of type "hdviz"
}
\examples{
some_viz <- hdviz(viz)
is_hdviz(some_viz)

}
